refact: convert nullvalue as enum
